Update struct of Intl static generation [major]
