+++ src/components/Collapse/index.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import CheatsheetDemo from '../CheatsheetDemo';
-import Icon from '../Icon';
+import CheatsheetDemo from 'src/components/CheatsheetDemo';
+import Icon from 'src/components/Icon';
 
 import * as styles from './Collapse.module.css';
