+++ src/components/Footer/index.jsx
@@ -1,5 +1,5 @@
 import { FormattedMessage } from 'react-intl';
-import Icon from '../Icon';
+import Icon from 'src/components/Icon';
 
 import * as styles from './Footer.module.css';
