+++ src/localization/en/cheatsheet.json
