+++ src/shortcuts.js
@@ -4,14 +4,10 @@ const shortcuts = {
   nextStep: "enter",
   prevStep: "shift+enter",
   hint: "alt+h",
-  languageSwitch: "alt+l",
   flagGlobal: "alt+g",
   flagMultiline: "alt+m",
   flagCaseInsenstive: "alt+c",
   close: "esc",
-  up: "up",
-  down: "down",
-  languageSelect: "space",
 };
 
 export default shortcuts;
